% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cousinsDemo}
\alias{cousinsDemo}
\title{Example dataset with autosomal SNPs for two related individuals}
\format{
A tibble with 3,915 rows and 9 variables:
\itemize{
\item \code{CHROM}: Chromosome label
\item \code{MARKER}: SNP identifier
\item \code{MB}: Physical position in megabases
\item \code{CM}: Map position in centiMorgans
\item \code{A1}: First SNP allele
\item \code{A2}: Second SNP allele
\item \code{FREQ1}: Population frequency of \code{A1}
\item \code{ID1}: Genotype of individual 1
\item \code{ID2}: Genotype of individual 2
}
}
\usage{
cousinsDemo
}
\description{
A dataset with 3,915 autosomal SNPs (from the FORCE panel) typed in two
related individuals.
}
\examples{
cousinsDemo
}
\keyword{datasets}
