% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findSegments.R
\name{findSegments}
\alias{findSegments}
\title{Identify IBD segments}
\usage{
findSegments(data, ids = NULL, k1, a)
}
\arguments{
\item{data}{Data frame with required columns \code{chrom}, \code{cm}, \code{a1} and \code{freq1}.}

\item{ids}{Genotype columns (default: last 2 columns).}

\item{k1, a}{HMM parameters.}
}
\value{
Data frame with IBD segments.
}
\description{
Identify IBD segments
}
\examples{
findSegments(cousinsDemo, k1 = 0.2, a = 5)

}
\seealso{
\code{\link[=plotIBD]{plotIBD()}}
}
