% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitHMM.R
\name{fitHMM}
\alias{fitHMM}
\title{Fit a Hidden Markov Model to genotype data}
\usage{
fitHMM(
  data,
  ids = NULL,
  k1 = NULL,
  a = NULL,
  thompson = TRUE,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{data}{Data frame with required columns \code{chrom}, \code{cm}, \code{a1} and \code{freq1}
(case insensitive).}

\item{ids}{Genotype columns (default: last 2 columns).}

\item{k1, a}{Numeric HMM parameters. Supplying a value fixes the parameter; if
NULL (default), the parameter is estimated.}

\item{thompson}{A logical indicating the optimisation method. (See Details.)}

\item{verbose}{A logical indicating whether to print information during the
optimisation.}

\item{...}{Additional arguments passed to the \code{control} parameter of
\code{\link[stats:optim]{stats::optim()}}.}
}
\value{
A list containing the fitted parameters \code{k1} and \code{a}, and some
additional information about the optimisation.
}
\description{
This function fits a continuous-time HMM to the given genotype data. It does
so by optimising the parameters \code{k1} (the probability of being in an IBD
state) and the transition rate \code{a} to maximise the total log-likelihood.
}
\details{
If \code{thompson = TRUE} (default), the parameter \code{k1} is estimated first, using
the standard maximum-likelihood approach for estimating pairwise relatedness
coefficients (first described by Thompson 1975). The actual work is done by a
call to \code{\link[forrel:ibdEstimate]{forrel::ibdEstimate()}}. The parameter \code{a} is subsequently estimated
with \code{optim}.

If \code{thompson = FALSE}, both parameters \code{k1} and \code{a} are optimised together.
(This is the only method for X chromosomal data.)
}
\examples{
\dontrun{
fitHMM(cousinsDemo)
}

}
\seealso{
\code{\link[=totalLoglik]{totalLoglik()}}, \code{\link[forrel:ibdEstimate]{forrel::ibdEstimate()}}
}
