% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ibdfindr-package.R
\docType{package}
\name{ibdfindr-package}
\alias{ibdfindr}
\alias{ibdfindr-package}
\title{ibdfindr: HMM Toolkit for Inferring IBD Segments from SNP Genotypes}
\description{
Implements continuous-time hidden Markov models to infer identity-by-descent (IBD) segments shared by two individuals from their SNP genotypes. Provides posterior probabilities at each marker (forward-backward), prediction of IBD segments (Viterbi), and functions for visualising results. Supports both autosomal and X chromosomes.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/magnusdv/ibdfindr}
  \item Report bugs at \url{https://github.com/magnusdv/ibdfindr/issues}
}

}
\author{
\strong{Maintainer}: Magnus Dehli Vigeland \email{m.d.vigeland@medisin.uio.no} (\href{https://orcid.org/0000-0002-9134-4962}{ORCID})

}
\keyword{internal}
