% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findSegments.R
\name{findSegments}
\alias{findSegments}
\title{Identify IBD segments}
\usage{
findSegments(data, ids = NULL, k1, a, prepped = FALSE, verbose = FALSE)
}
\arguments{
\item{data}{Data frame with required columns \code{chrom}, \code{cm}, \code{a1} and \code{freq1}.}

\item{ids}{Genotype columns (default: last 2 columns).}

\item{k1, a}{HMM parameters. See \code{\link[=fitHMM]{fitHMM()}} for how to estimate these.}

\item{prepped}{A logical indicating if the input data has undergone internal
prepping. Can be ignored by most users.}

\item{verbose}{A logical.}
}
\value{
Data frame with IBD segments.
}
\description{
Identifies genomic segments shared identical-by-descent (IBD) between two
individuals from SNP marker data. The method applies a hidden Markov model
(HMM) along each chromosome, with states 0 (non-IBD) and 1 (IBD), and uses
the Viterbi algorithm to infer the most likely sequence of states.
}
\examples{
findSegments(cousinsDemo, k1 = 0.2, a = 5)

}
\seealso{
\code{\link[=plotIBD]{plotIBD()}}
}
