% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ibdPosteriors.R
\name{ibdPosteriors}
\alias{ibdPosteriors}
\title{IBD posteriors}
\usage{
ibdPosteriors(data, ids = NULL, k1, a)
}
\arguments{
\item{data}{Data frame with required columns \code{chrom}, \code{cm}, \code{a1} and \code{freq1}.}

\item{ids}{Genotype columns (default: last 2 columns).}

\item{k1, a}{HMM parameters.}
}
\value{
Data frame similar to \code{data}, with a column \code{post} containing the
posterior IBD probability at each marker locus.
}
\description{
IBD posteriors
}
\examples{
ibdPosteriors(cousinsDemo, k1 = 0.2, a = 5)

}
\seealso{
\code{\link[=plotIBD]{plotIBD()}}
}
