% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotIBD.R
\name{plotIBD}
\alias{plotIBD}
\title{Plot IBD segments and posteriors}
\usage{
plotIBD(data, segments = NULL, chrom = NULL, ncol = NULL)
}
\arguments{
\item{data}{A data frame, typically produced by \code{\link[=ibdPosteriors]{ibdPosteriors()}}.}

\item{segments}{A data frame with IBD segments, typically produced by \code{\link[=findSegments]{findSegments()}}.}

\item{chrom}{A vector of chromosomes to plot (default: all).}

\item{ncol}{Number of columns in the plot. By default a suitable layout
is chosen automatically.}
}
\value{
A \code{ggplot2} plot.
}
\description{
Plot IBD segments and posteriors
}
\examples{
x = subset(cousinsDemo, CHROM \%in\% 11:13)

post = ibdPosteriors(x, k1 = 0.2, a = 10)
segs = findSegments(x, k1 = 0.2, a = 10)
plotIBD(post, segs)

}
\seealso{
\code{\link[=findSegments]{findSegments()}}, \code{\link[=ibdPosteriors]{ibdPosteriors()}}
}
