% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotIBD.R
\name{plotIBD}
\alias{plotIBD}
\title{Plot IBD segments and posteriors}
\usage{
plotIBD(x, segments = NULL, chrom = NULL, ncol = NULL)
}
\arguments{
\item{x}{A data frame, typically produced by \code{\link[=ibdPosteriors]{ibdPosteriors()}}, or a list
produced by \code{\link[=findIBD]{findIBD()}}.}

\item{segments}{A data frame with IBD segments, typically produced by
\code{\link[=findSegments]{findSegments()}}.}

\item{chrom}{A vector of chromosomes to plot (default: all).}

\item{ncol}{Number of columns in the plot. By default a suitable layout is
chosen automatically.}
}
\value{
A \code{ggplot2} plot.
}
\description{
Plot IBD segments and posteriors
}
\examples{
x = subset(cousinsDemo, CHROM \%in\% 3:4)
ibd = findIBD(x, k1 = 0.2, a = 5)
plotIBD(ibd)

}
\seealso{
\code{\link[=findSegments]{findSegments()}}, \code{\link[=ibdPosteriors]{ibdPosteriors()}}
}
